Building a basic conversation
This is my fifth submit:
The conversation object containing CharacterName and ConversationText and DisplayPic has been saving and serializing in the Project and can be created in Editor.
The Conversation Manager is written with the Singleton.cs from Unity Wiki.
And the UI displaying the conversation elements is created to be finished.