Setting the battle scene and saving the map position
This is my 9th submit:
The battle scene was set up with simple state machines and BattleManager. And the run away button was added.
The battle was started with random enmeies and a  probability in two different area.
And the last map position in town and overworld can be saved with GameState.
And the build exe was removed.